feat: Overhaul UI, fix bugs, and add features
This commit addresses a wide range of issues and feature requests to improve the Quran application.

- **Settings Page Overhaul:**
  - Replaced the settings modal with a dedicated, full-screen settings page (`src/pages/SettingsPage.tsx`).
  - Removed the "Performance Settings" and "Font Size" controls.
  - Replaced all emojis with `lucide-react` icons for a cleaner and more consistent UI.
  - Reorganized the settings layout for better usability.

- **Bug Fixes:**
  - Fixed the "Show Translation" and "Show Tafsir" toggles, which now correctly update the view.
  - Resolved the "Error playing audio" bug by correcting the audio URL construction logic.

- **UI/UX Enhancements:**
  - Redesigned the audio player with a modern UI, including a new progress bar, volume controls, and icons.
  - Implemented verse highlighting to visually mark the currently playing verse.

- **Code Quality:**
  - Refactored state management for settings and audio playback to be more robust and centralized in the `QuranReader` component.
  - Removed unused code and variables.
